+++ internal/discord/commands.go
@@ -1,0 +1,5 @@
+package discord
+
+// Possibly for later:
+// - Set mention in db to active
+// - Set mention in db to inactive
