+++ internal/discord/events.go
@@ -1,4 +1,4 @@
-package internal
+package discord
 
 import (
 	"log"
@@ -6,20 +6,18 @@ import (
 	dg "github.com/bwmarrin/discordgo"
 )
 
-var counter = 1
-
 func (b *Bot) onMessage(s *dg.Session, m *dg.MessageCreate) {
 	if m.Author.ID == s.State.User.ID {
 		return
 	}
 	// If the message is "ping" reply with "Pong!"
-	if m.Content == "ping" {
-		s.ChannelMessageSend(m.ChannelID, "Pong!")
+	if m.Content == "ping" && m.ChannelID != b.config.ChannelID {
+		s.ChannelMessageSend(b.config.ChannelID, "Pong!")
 	}
 
 	// If the message is "pong" reply with "Ping!"
-	if m.Content == "pong" {
-		s.ChannelMessageSend(m.ChannelID, "Ping!")
+	if m.Content == "pong" && m.ChannelID != b.config.ChannelID {
+		s.ChannelMessageSend(b.config.ChannelID, "Ping!")
 	}
 }
